+++ src/elite/libs/EliteSharpLib/Views/SixteenBit/Intro1View16Bit.cs
@@ -16,7 +16,7 @@ internal sealed class Intro1View16Bit : BaseView16Bit, IView<Intro1Model>
     private const float CreditSpacing = 20;
 
     private readonly IEliteDraw _draw;
-    private readonly uint _colorGold;
+    private readonly uint _colorGoldenrod;
     private readonly uint _colorWhite;
 
     internal Intro1View16Bit(IEliteDraw draw)
@@ -24,7 +24,7 @@ internal sealed class Intro1View16Bit : BaseView16Bit, IView<Intro1Model>
     {
         _draw = draw;
 
-        _colorGold = draw.Palette["Gold"];
+        _colorGoldenrod = draw.Palette["Goldenrod"];
         _colorWhite = draw.Palette["White"];
     }
 
@@ -41,6 +41,6 @@ internal sealed class Intro1View16Bit : BaseView16Bit, IView<Intro1Model>
             y += CreditSpacing;
         }
 
-        _draw.Graphics.DrawTextCentre(_draw.Layout.ScannerTop - 30, model.Prompt, nameof(FontType.Large), _colorGold);
+        _draw.Graphics.DrawTextCentre(_draw.Layout.ScannerTop - 30, model.Prompt, nameof(FontType.Large), _colorGoldenrod);
     }
 }